ignite/backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv
psycopg2-binary